requirements.txt
opencv-python
mysql-connector-python
deepface
geopy
twilio
